requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
qdrant-client==1.17.1
langchain-classic==1.0.3
langchain-core==1.2.21
langchain-qdrant==1.1.0
langchain-community==0.4.1
sentence-transformers==5.3.0
litellm==1.82.6
google-cloud-aiplatform==1.142.0
google-auth==2.49.1
pydantic==2.12.5
python-dotenv==1.2.2
httpx==0.28.1
pytest==9.0.2
